+++ system/modules/isotope/CHANGELOG.md
@@ -39,6 +39,7 @@ Version 1.3.rc2 (2012-??-??)
 - trim extra backslash from URLs returned in Products Search Indexing
 - The sorting for Order ID was not correct (#2871)
 - Duplicating an attribute caused an empty field to be added in database.sql (#2955)
+- saveNcreate did not work in ModuleIsotopeSetup (#2933)
 
 ### Other
 - Introducing the "provider" directory for complex callback handlers